refactor: move buzzer box components to new folder
