+++ package.json
@@ -15,7 +15,7 @@
     "@dnd-kit/core": "^6.1.0",
     "@dnd-kit/sortable": "^8.0.0",
     "@dnd-kit/utilities": "^3.2.2",
-    "@prisma/client": "^5.14.0",
+    "@prisma/client": "^5.15.0",
     "@radix-ui/react-select": "^2.0.0",
     "@react-spring/web": "^9.7.3",
     "@tailwindcss/typography": "^0.5.12",
@@ -49,7 +49,7 @@
     "next-plugin-modernizr": "^0.0.3",
     "node-fetch": "^2.6.7",
     "postcss": "^8.4.31",
-    "prisma": "^5.14.0",
+    "prisma": "^5.15.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-error-boundary": "^4.0.11",
