+++ src/components/buzzer/box/TeamDisplay.tsx
